# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateScheduleIn": {
        "additionalProperties": false,
        "properties": {
          "agent_kind": {
            "description": "Tipo de agente.",
            "examples": [
              "site_summary"
            ],
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Kind",
            "type": "string"
          },
          "cadence": {
            "enum": [
              "daily",
              "weekly"
            ],
            "maxLength": 16,
            "title": "Cadence",
            "type": "string"
          },
          "site_id": {
            "description": "UUID del sitio.",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Site Id",
            "type": "string"
          },
          "time_local": {
            "examples": [
              "09:00"
            ],
            "maxLength": 5,
            "title": "Time Local",
            "type": "string"
          },
          "weekday": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekday"
          }
        },
        "required": [
          "agent_kind",
          "site_id",
          "cadence",
          "time_local"
        ],
        "title": "CreateScheduleIn",
        "type": "object"
      },
...
      "ErrorOut": {
        "additionalProperties": false,
        "description": "Formato de error común a todas las capas (ADR 0002).",
        "properties": {
          "code": {
            "description": "Código estable `dominio.motivo`.",
            "examples": [
              "engine.unauthorized"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Mensaje en español para el usuario.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScheduleListOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScheduleOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "null"
          }
        },
        "required": [
          "items"
        ],
        "title": "ScheduleListOut",
        "type": "object"
      },
      "ScheduleOut": {
        "additionalProperties": false,
        "properties": {
          "agent_kind": {
            "title": "Agent Kind",
            "type": "string"
          },
          "cadence": {
            "enum": [
              "daily",
              "weekly"
            ],
            "title": "Cadence",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "last_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Id"
          },
          "next_run_at": {
            "title": "Next Run At",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "time_local": {
            "examples": [
              "09:00"
            ],
            "title": "Time Local",
            "type": "string"
          },
          "timezone": {
            "description": "Zona IANA del sistema al guardarla.",
            "title": "Timezone",
            "type": "string"
          },
          "weekday": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "0 = lunes … 6 = domingo; solo semanales.",
            "title": "Weekday"
          }
        },
        "required": [
          "id",
          "agent_kind",
          "site_id",
          "cadence",
          "weekday",
          "time_local",
          "timezone",
          "enabled",
          "next_run_at",
          "last_run_at",
          "last_run_id"
        ],
        "title": "ScheduleOut",
        "type": "object"
      },
...
      "UpdateScheduleIn": {
        "additionalProperties": false,
        "properties": {
          "cadence": {
            "anyOf": [
              {
                "enum": [
                  "daily",
                  "weekly"
                ],
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cadence"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "time_local": {
            "anyOf": [
              {
                "maxLength": 5,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Local"
          },
          "weekday": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekday"
          }
        },
        "title": "UpdateScheduleIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Faro Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/schedules": {
      "get": {
        "operationId": "listSchedules",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Programaciones de los agentes",
        "tags": [
          "schedules"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      },
      "post": {
        "operationId": "createSchedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.unknown` o `site.not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`schedule.duplicate`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`schedule.invalid` o `agent.site_required`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Programar un agente en un sitio (cada día o cada semana)",
        "tags": [
          "schedules"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/schedules/{schedule_id}": {
      "delete": {
        "operationId": "deleteSchedule",
        "parameters": [
          {
            "description": "UUID de la programación en minúsculas.",
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "description": "UUID de la programación en minúsculas.",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`schedule.not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Quitar una programación",
        "tags": [
          "schedules"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      },
      "patch": {
        "operationId": "updateSchedule",
        "parameters": [
          {
            "description": "UUID de la programación en minúsculas.",
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "description": "UUID de la programación en minúsculas.",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`schedule.not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`schedule.invalid`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Activar, desactivar o cambiar la frecuencia o la hora",
        "tags": [
          "schedules"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
...
  }
}